perf(ycbcr): use fixed-point for YCbCr conversions
Converts YCbCr components from f32 to u8.
Implements RGB-YCbCr conversions using 16-bit fixed-point
arithmetic for improved performance and memory usage.
Adjusts test tolerance due to integer rounding.

+++ src/color/Ycbcr.zig
@@ -20,11 +20,11 @@ const Xyb = @import("Xyb.zig");
 const Xyz = @import("Xyz.zig");
 
 /// Y component (luma/brightness) in range [0, 255]
-y: f32,
+y: u8,
 /// Cb component (blue-difference chroma) in range [0, 255] (128 = neutral)
-cb: f32,
+cb: u8,
 /// Cr component (red-difference chroma) in range [0, 255] (128 = neutral)
-cr: f32,
+cr: u8,
 
 const Ycbcr = @This();
